refactor: move fetchCounts/ref logic to composable
